+++ src/main/java/com/audiolibrary/web/controller/ArtistController.java
@@ -5,6 +5,7 @@ import com.audiolibrary.web.model.Album;
 import com.audiolibrary.web.model.Artist;
 import com.audiolibrary.web.repository.AlbumRepository;
 import com.audiolibrary.web.repository.ArtistRepository;
+import com.audiolibrary.web.service.ArtistService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.dao.DataIntegrityViolationException;
@@ -25,50 +26,12 @@ import java.util.Optional;
 public class ArtistController {
 
     @Autowired
-    private ArtistRepository artistRepository;
+    private ArtistService artistService;
 
-    @Autowired
-    private AlbumRepository albumRepository;
-    @Value("${deleteAlbumsWithArtist}")
-    private Boolean DELETE_ALBUMS_WITH_ARTIST;
-
-    private void ExceptionForPagination(String name, Integer page, Integer size, Sort.Direction sortDirection, String sortProperty){
-        List<String> sortPropertyList = Arrays.asList("name", "id");
-        int maxPage;
-        if(!sortPropertyList.contains(sortProperty)){
-            throw new IllegalArgumentException("Le tri fonctionne uniquement sur 'id' et 'name'");
-        }
-
-        if (size < 2 || size > 50)  {
-            throw new IllegalArgumentException("La taille des pages doit être comprise entre 1 et 50 !");
-        }
-
-        if (name != null){
-            if(page < 0){
-                throw new IllegalArgumentException("Le numéro de page ne peut être inférieur à 0");
-            }
-            maxPage = artistRepository.findAllByNameContaining(name, PageRequest.of(
-                    page,
-                    size,
-                    sortDirection,
-                    sortProperty)
-            ).getTotalPages() - 1;
-        }else{
-            maxPage = (int) artistRepository.count() / size - 1;
-        }
-
-        if (page < 0 || page > maxPage){
-            throw new IllegalArgumentException("Le numéro de page ne peut être inférieur à 0 ou supérieur à " + maxPage);
-        }
-    }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
     public Artist getArtistById(@PathVariable("id") Integer id){
-        Optional<Artist> artist = artistRepository.findById(id);
-        if(artist.isPresent()) {
-            return artist.get();
-        }
-        throw new EntityNotFoundException("L'artiste d'identifiant " + id + " n'existe pas.");
+        return artistService.findById(id);
     }
 
     @RequestMapping(value = "", params = {"name"})
@@ -79,9 +42,7 @@ public class ArtistController {
             @RequestParam(value = "sortProperty", defaultValue = "name") String sortProperty,
             @RequestParam(value = "sortDirection", defaultValue = "ASC") Sort.Direction sortDirection){
 
-        ExceptionForPagination(name, page, size, sortDirection, sortProperty);
-
-        return artistRepository.findAllByNameContaining(name, PageRequest.of(page, size, sortDirection, sortProperty));
+        return artistService.findMyName(name, page, size, sortDirection, sortProperty);
     }
 
     @GetMapping
@@ -91,9 +52,7 @@ public class ArtistController {
             @RequestParam(value = "sortProperty", defaultValue = "name") String sortProperty,
             @RequestParam(value = "sortDirection", defaultValue = "ASC") Sort.Direction sortDirection){
 
-        ExceptionForPagination(null, page, size, sortDirection, sortProperty);
-
-        return artistRepository.findAll(PageRequest.of(page, size, sortDirection, sortProperty));
+        return artistService.findAllArtists(page, size, sortDirection, sortProperty);
     }
 
     @RequestMapping(
@@ -102,19 +61,7 @@ public class ArtistController {
             produces = MediaType.APPLICATION_JSON_VALUE
     )
     public Artist createArtist(@RequestBody Artist artist) {
-        if (artist.getName() == null) {
-            throw new IllegalStateException ("L'artiste ne peut avoir un nom 'null' !");
-        }
-
-        if (artist.getName().trim().isEmpty()) {
-            throw new IllegalStateException ("L'artiste ne peut avoir un nom vide !");
-        }
-
-        if(artistRepository.existsByName(artist.getName().trim())) {
-            throw new DataIntegrityViolationException("L'artiste avec le nom : " + artist.getName() + ", existe déjà !");
-        }
-
-        return artistRepository.save(artist);
+        return artistService.createArtist(artist);
     }
 
     @RequestMapping(
@@ -123,50 +70,14 @@ public class ArtistController {
             consumes = MediaType.APPLICATION_JSON_VALUE,
             produces = MediaType.APPLICATION_JSON_VALUE
     )
-
     public Artist updateArtist(@PathVariable("id") Integer id, @RequestBody Artist artist){
-        if (artist.getName().trim().isEmpty()) {
-            throw new IllegalStateException ("L'artiste doit avoir un nom !");
-        }
-
-        if(!artistRepository.existsById(id)) {
-            throw new EntityNotFoundException("L'artiste avec l'id : " + id + ", n'existe pas !");
-        }
-
-        if(!id.equals(artist.getId())) {
-            throw new IllegalArgumentException("L'id ne correspond pas à celui de cet artiste");
-        }
-
-        if(artistRepository.existsByName(artist.getName().trim())){
-            throw new DataIntegrityViolationException("L'artiste existe déjà !");
-        }
-
-        return artistRepository.save(artist);
+        return artistService.updateArtist(id, artist);
     }
 
     @RequestMapping(method = RequestMethod.DELETE, value = "/{id}")
     @ResponseStatus(value = HttpStatus.NO_CONTENT)
     public void deleteArtist(@PathVariable ("id") Integer id){
-        Optional<Artist> artistOptional = artistRepository.findById(id);
-
-        if(artistOptional.isEmpty()) {
-            throw new EntityNotFoundException("L'artiste avec l'id : " + id + ", n'existe pas !");
-        }
-
-        Artist artist = artistOptional.get();
-
-        if(!artist.getAlbums().isEmpty()) {
-            if (DELETE_ALBUMS_WITH_ARTIST) {
-                List<Album> albums = artist.getAlbums();
-                //CASCADE ?
-                for (Album album : albums) {
-                    albumRepository.delete(album);
-                }
-            }else {
-                throw new DataIntegrityViolationException("Cet artiste dispose d'albums, vous ne pouvez pas le supprimer !");
-            }
-        }
-        artistRepository.deleteById(id);
+        artistService.deleteArtist(id);
     }
 
 }
